make helper classes an inner class of the word
